more algorithms

+++ URI/1002.py
@@ -1,4 +1,4 @@
 n = 3.14159
-raio = float(input())
+raio = int(input())
 area = n * raio ** 2
 print('A={:.4f}'.format(area))
