+++ URI/1005.py
@@ -1,4 +1,4 @@
-A = float(input())
-B = float(input())
+A = int(input())
+B = int(input())
 MEDIA = ((A* 3.5) + (B * 7.5))/11
 print('MEDIA = {:.5f}'.format(MEDIA))
