+++ URI/1006.py
@@ -1,5 +1,5 @@
-A = float(input())
-B = float(input())
-C = float(input())
+A = int(input())
+B = int(input())
+C = int(input())
 MEDIA = ((A * 2) + (B * 3) + (C * 5))/10
 print('MEDIA = {:.1f}'.format(MEDIA))
