+++ URI/1008.py
@@ -1,6 +1,6 @@
 NUMBER = int(input())
 b = int(input())
-c = float(input())
+c = int(input())
 SALARY = b * c
 print('NUMBER = {}'.format(NUMBER))
 print('SALARY = U$ {:.2f}'.format(SALARY))
